bot/requirements.txt
aiogram==3.19.0
python-dotenv==1.0.1
pycryptodome==3.22.0 